requirements.txt
Django~=2.2.11
django-guardian~=2.2.0
django-simple-captcha~=0.5.12
djangorestframework

https://codeload.github.com/sshwsfc/xadmin/zip/django2
